+++ packages/block-editor/src/hooks/duotone.jsx
@@ -233,7 +233,7 @@ function addDuotoneAttributes( settings ) {
 
 	// Allow blocks to specify their own attribute definition with default
 	// values if needed.
-	if ( ! settings.attributes.style ) {
+	if ( ! settings.attributes?.style ) {
 		Object.assign( settings.attributes, {
 			style: {
 				type: 'object',