+++ packages/block-editor/src/hooks/style.jsx
@@ -544,7 +544,7 @@ function addAttribute( settings ) {
 	}
 
 	// Allow blocks to specify their own attribute definition with default values if needed.
-	if ( ! settings.attributes.style ) {
+	if ( ! settings.attributes?.style ) {
 		Object.assign( settings.attributes, {
 			style: {
 				type: 'object',